primeira imagem renderizada com funções de raycast que descobrem intersecção.

+++ Vec2.h
@@ -1,9 +1,9 @@
 #pragma once
 
 struct Vector2{
-  int x, y;
+  float x, y;
 
-  constexpr Vector2(int x, int y) : x(x), y(y) {};
+  constexpr Vector2(float x, float y) : x(x), y(y) {};
 
   constexpr Vector2(unsigned x, unsigned y) : 
     x(static_cast<int>(x)), y(static_cast<int>(y)) {}
